Post Authentication handler for password expiry feature with unit tests.
