--- component/authenticator1/src/test/resources/testng.xml
@@ -1,9 +1,0 @@
-<!DOCTYPE suite SYSTEM "http://testng.org/testng-1.0.dtd" >
-
-<suite name="Identity-connector-password-policy-test-suite">
-    <test name="password-policy-authenticator-all" parallel="false">
-        <classes>
-            <class name="org.wso2.carbon.extension.identity.handler.passwordpolicy.test.PasswordPolicyUtilsTest"/>
-        </classes>
-    </test>
-</suite>
