+++ component/handler/src/main/java/org/wso2/carbon/identity/post/authn/handler/resetPasswordEnforcer/PasswordPolicyConstants.java
@@ -1,5 +1,5 @@
 /*
- *  Copyright (c) 2018, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
+ *  Copyright (c) 2023, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
  *
  *  WSO2 Inc. licenses this file to you under the Apache License,
  *  Version 2.0 (the "License"); you may not use this file except
@@ -17,67 +17,93 @@
  *
  */
 
-package org.wso2.carbon.identity.policy.password;
+package org.wso2.carbon.identity.post.authn.handler.resetPasswordEnforcer;
 
 /**
  * Password Change authenticator's constants
  */
 public class PasswordPolicyConstants {
-    public static final String AUTHENTICATOR_NAME = "password-reset-enforcer";
-    public static final String AUTHENTICATOR_FRIENDLY_NAME = "Password Reset Enforcer";
     public static final String AUTHENTICATOR_TYPE = "LOCAL";
-    public static final String STATE = "state";
-
-    public static final String CURRENT_PWD = "CURRENT_PWD";
-    public static final String NEW_PWD = "NEW_PWD";
-    public static final String NEW_PWD_CONFIRMATION = "NEW_PWD_CONFIRMATION";
-
     public static final String LAST_CREDENTIAL_UPDATE_TIMESTAMP_CLAIM =
             "http://wso2.org/claims/identity/lastPasswordUpdateTime";
-
     public static final String LAST_CREDENTIAL_UPDATE_TIMESTAMP_CLAIM_NON_IDENTITY =
             "http://wso2.org/claims/lastPasswordChangedTimestamp";
-
     public static final String PASSWORD_CHANGE_STREAM_NAME =
             "org.wso2.carbon.identity.policy.password.PendingNotifications:1.0.0";
     public static final String PASSWORD_CHANGE_EVENT_HANDLER_NAME = "passwordExpiry";
-
-    public static final String CONNECTOR_CONFIG_FRIENDLY_NAME = "Password Expiry";
-    public static final String CONNECTOR_CONFIG_CATEGORY = "Password Policies";
-    public static final String CONNECTOR_CONFIG_SUB_CATEGORY = "DEFAULT";
-
     public static final String CONNECTOR_CONFIG_PASSWORD_EXPIRY_IN_DAYS = "passwordExpiry.passwordExpiryInDays";
     public static final String CONNECTOR_CONFIG_PASSWORD_EXPIRY_IN_DAYS_DISPLAYED_NAME = "Password Expiry In Days";
     public static final String CONNECTOR_CONFIG_PASSWORD_EXPIRY_IN_DAYS_DESCRIPTION =
             "Number of days after which the password will expire";
     public static final int CONNECTOR_CONFIG_PASSWORD_EXPIRY_IN_DAYS_DEFAULT_VALUE = 30;
-
     public static final String CONNECTOR_CONFIG_ENABLE_EMAIL_NOTIFICATIONS = "passwordExpiry.enableEmailNotifications";
+    public static final String CONNECTOR_CONFIG_ENABLE_PASSWORD_EXPIRY = "passwordExpiry.enablePasswordExpiry";
+    public static final String CONNECTOR_CONFIG_ENABLE_PASSWORD_EXPIRY_DISPLAYED_NAME =
+            "Enable Password Expiry";
     public static final String CONNECTOR_CONFIG_ENABLE_EMAIL_NOTIFICATIONS_DISPLAYED_NAME =
             "Enable Sending Email Notifications";
-    public static final String CONNECTOR_CONFIG_ENABLE_EMAIL_NOTIFICATIONS_DESCRIPTION =
-            "Enable to send email notifications to reset the password. " +
-                    "This requires an IS Analytics instance.";
-    public static final boolean CONNECTOR_CONFIG_ENABLE_EMAIL_NOTIFICATIONS_DEFAULT_VALUE = false;
-
     public static final String CONNECTOR_CONFIG_PRIOR_REMINDER_TIME_IN_DAYS = "passwordExpiry.priorReminderTimeInDays";
     public static final String CONNECTOR_CONFIG_PRIOR_REMINDER_TIME_IN_DAYS_DISPLAYED_NAME =
             "Prior Reminder Time In Days";
+    public static final String CONNECTOR_CONFIG_ENABLE_EMAIL_NOTIFICATIONS_DESCRIPTION =
+            "Enable to send email notifications to reset the password. " +
+                    "This requires an IS Analytics instance.";
     public static final String CONNECTOR_CONFIG_PRIOR_REMINDER_TIME_IN_DAYS_DESCRIPTION =
             "Number of days before the password expiry that the users should be reminded of password expiry";
+    public static final String CONNECTOR_CONFIG_SUB_CATEGORY = "DEFAULT";
+    public static final boolean CONNECTOR_CONFIG_ENABLE_EMAIL_NOTIFICATIONS_DEFAULT_VALUE = false;
     public static final int CONNECTOR_CONFIG_PRIOR_NOTICE_TIME_IN_DAYS_DEFAULT_VALUE = 0;
-
-    public static final String LOGIN_STANDARD_PAGE = "login.do";
-    public static final String PASSWORD_RESET_ENFORCER_PAGE = "pwd-reset.jsp";
-    public static final String PASSWORD_HISTORY_VIOLATION_ERROR_CODE = "22001";
-    public static final String LAST_FAILED_AUTHENTICATOR = "LastFailedAuthenticator";
-
-    //password grant flow - password expiry validation constants
+    public static final String PASSWORD_EXPIRY_IN_DAYS_DEFAULT_VALUE = "30";
+    public static final String PASSWORD_EXPIRED_ERROR_MESSAGE = "Password has expired";
+    public static final String CONNECTOR_CONFIG_FRIENDLY_NAME = "Password Expiry";
+    public static final String CONNECTOR_CONFIG_CATEGORY = "Password Policies";
     public static final String PASSWORD_GRANT_POST_AUTHENTICATION_EVENT = "PASSWORD_GRANT_POST_AUTHENTICATION";
     public static final String AUTHENTICATION_STATUS = "authenticationStatus";
-    public static final String PASSWORD_EXPIRED_ERROR_MESSAGE = "Password has expired";
-    public static final String PASSWORD_EXPIRY_IN_DAYS_DEFAULT_VALUE = "30";
+    public static final String BASIC_AUTHENTICATOR = "BasicAuthenticator";
+    public static final String ENFORCE_PASSWORD_RESET_HANDLER = "EnforcePasswordResetHandler";
+    public static final String FALSE = "false";
 
-    private PasswordPolicyConstants() {      // To prevent instantiation
+    public enum ErrorMessages {
+        ERROR_WHILE_GETTING_FEDERATED_USERNAME("80027", "Error while getting the federated username"),
+        ERROR_WHILE_GETTING_USER_STORE_DOMAIN("80020", "Error occurred while getting the user store domain."),
+        ERROR_WHILE_GETTING_USER_REALM("80029", "Error occurred while getting the user realm."),
+        ERROR_WHILE_GETTING_CLAIM_MAPPINGS("80013", "Error while getting claim mappings for user, %s"),
+        ERROR_WHILE_READING_SYSTEM_CONFIGURATIONS("80028", "Error while reading the configurations.");
+        private final String code;
+        private final String message;
+
+        /**
+         * Create an Error Message.
+         *
+         * @param code    Relevant error code.
+         * @param message Relevant error message.
+         */
+        ErrorMessages(String code, String message) {
+            this.code = code;
+            this.message = message;
+        }
+
+        /**
+         * To get the code of specific error.
+         *
+         * @return Error code.
+         */
+        public String getCode() {
+            return code;
+        }
+
+        /**
+         * To get the message of specific error.
+         *
+         * @return Error message.
+         */
+        public String getMessage() {
+            return message;
+        }
+
+        @Override
+        public String toString() {
+            return code + " - " + message;
+        }
     }
 }
