+++ component/handler/src/main/java/org/wso2/carbon/identity/post/authn/handler/resetPasswordEnforcer/internal/EnforcePasswordResetPostAuthenticationHandlerDataHolder.java
@@ -1,0 +1,57 @@
+/*
+ * Copyright (c) 2023, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
+ *
+ * WSO2 Inc. licenses this file to you under the Apache License,
+ * Version 2.0 (the "License"); you may not use this file except
+ * in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+
+package org.wso2.carbon.identity.post.authn.handler.resetPasswordEnforcer.internal;
+
+import org.osgi.framework.BundleContext;
+import org.wso2.carbon.user.core.service.RealmService;
+
+/**
+ * DataHolder for count users and roles in user store.
+ */
+public class EnforcePasswordResetPostAuthenticationHandlerDataHolder {
+    private static EnforcePasswordResetPostAuthenticationHandlerDataHolder instance
+            = new EnforcePasswordResetPostAuthenticationHandlerDataHolder();
+    private BundleContext bundleContext = null;
+    private RealmService realmService = null;
+
+    private EnforcePasswordResetPostAuthenticationHandlerDataHolder() {
+
+    }
+
+    public static EnforcePasswordResetPostAuthenticationHandlerDataHolder getInstance() {
+        return instance;
+    }
+
+    public RealmService getRealmService() {
+        return realmService;
+    }
+
+    public void setRealmService(RealmService realmService) {
+        this.realmService = realmService;
+    }
+
+    public BundleContext getBundleContext() {
+        return bundleContext;
+    }
+
+    public void setBundleContext(BundleContext bundleContext) {
+        this.bundleContext = bundleContext;
+    }
+
+}
