+++ component/handler/src/test/java/org/wso2/carbon/extension/identity/handler/passwordpolicy/test/PasswordPolicyTestConstants.java
@@ -1,0 +1,36 @@
+/*
+ *  Copyright (c) 2023, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
+ *
+ *  WSO2 Inc. licenses this file to you under the Apache License,
+ *  Version 2.0 (the "License"); you may not use this file except
+ *  in compliance with the License.
+ *  You may obtain a copy of the License at
+ *
+ *  http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  Unless required by applicable law or agreed to in writing,
+ *  software distributed under the License is distributed on an
+ *  "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ *  KIND, either express or implied.  See the License for the
+ *  specific language governing permissions and limitations
+ *  under the License.
+ *
+ */
+
+package org.wso2.carbon.extension.identity.handler.passwordpolicy.test;
+
+/**
+ * Password Change authenticator's constants
+ */
+public class PasswordPolicyTestConstants {
+
+    public static final String LOCAL = "LOCAL";
+    public static final String IS_PASSWORD_EXPIRY_ENABLED = "isPasswordExpiryEnabled";
+    public static final String CARBON_SUPER = "carbon.super";
+    public static final String TRUE = "true";
+    public static final String HTTP_WSO_2_ORG_CLAIMS_IDENTITY_LAST_PASSWORD_UPDATE_TIME
+            = "http://wso2.org/claims/identity/lastPasswordUpdateTime";
+    public static final String HTTP_WSO_2_ORG_CLAIMS_IDENTITY_LAST_PASSWORD_UPDATE_TIME1
+            = "http://wso2.org/claims/identity/lastPasswordUpdateTime";
+    public static final String BOB = "bob";
+}
