+++ pom.xml
@@ -33,7 +33,6 @@
 
     <modules>
         <module>component/authenticator</module>
-        <module>component/authenticator1</module>
         <module>feature/org.wso2.carbon.extension.identity.authenticator.passwordpolicy.feature</module>
     </modules>
 
@@ -100,11 +99,6 @@
                 <version>${carbon.kernel.version}</version>
             </dependency>
             <dependency>
-                <groupId>org.wso2.carbon</groupId>
-                <artifactId>org.wso2.carbon.user.api</artifactId>
-                <version>${carbon.kernel.version}</version>
-            </dependency>
-            <dependency>
                 <groupId>org.wso2.carbon.identity.framework</groupId>
                 <artifactId>org.wso2.carbon.identity.application.mgt</artifactId>
                 <version>${carbon.identity.framework.version}</version>
@@ -159,12 +153,12 @@
                 <groupId>org.powermock</groupId>
                 <artifactId>powermock-api-mockito</artifactId>
                 <version>${powermock.version}</version>
-                <exclusions>
-                    <exclusion>
-                        <groupId>cglib</groupId>
-                        <artifactId>cglib-nodep</artifactId>
-                    </exclusion>
-                </exclusions>
+                <scope>test</scope>
+            </dependency>
+            <dependency>
+                <groupId>org.mockito</groupId>
+                <artifactId>mockito-all</artifactId>
+                <version>${mockito.version}</version>
                 <scope>test</scope>
             </dependency>
             <dependency>
@@ -173,19 +167,6 @@
                 <version>${powermock.version}</version>
                 <scope>test</scope>
             </dependency>
-           <!-- <dependency>
-                <groupId>org.powermock</groupId>
-                <artifactId>powermock-core</artifactId>
-                <version>${powermock.version}</version>
-                <scope>test</scope>
-            </dependency>
-            <dependency>
-                <groupId>org.powermock</groupId>
-                <artifactId>powermock-module-junit4</artifactId>
-                <version>${powermock.version}</version>
-                <scope>test</scope>
-            </dependency>-->
-
         </dependencies>
     </dependencyManagement>
 
@@ -339,14 +320,12 @@
         <carbon.identity.framework.version>5.11.109</carbon.identity.framework.version>
         <carbon.identity.governance.version>1.1.7</carbon.identity.governance.version>
         <commons-logging.version>4.4.23</commons-logging.version>
-        <carbon.kernel.version>4.7.1</carbon.kernel.version>
+        <carbon.kernel.version>4.4.23</carbon.kernel.version>
         <testng.version>6.9.10</testng.version>
         <jacoco.version>0.7.9</jacoco.version>
         <maven.surefire.plugin.version>2.18.1</maven.surefire.plugin.version>
-        <wso2.maven.compiler.source>1.8</wso2.maven.compiler.source>
-        <wso2.maven.compiler.target>1.8</wso2.maven.compiler.target>
-        <mockito.version>3.10.0</mockito.version>
-        <powermock.version>1.7.3</powermock.version>
+        <mockito.version>1.10.19</mockito.version>
+        <powermock.version>1.6.5</powermock.version>
         <maven.bundle.plugin.version>2.4.0</maven.bundle.plugin.version>
         <carbon.analytics-common.version>5.1.31</carbon.analytics-common.version>
         <maven.scr.plugin.version>1.24.0</maven.scr.plugin.version>
